+++ Sources/VCore/Extensions/SwiftUI/View.BlocksGestures.swift
@@ -1,5 +1,5 @@
 //
-//  View.OverlayGestureBlocker.swift
+//  View.BlocksGestures.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 03.03.23.
@@ -7,20 +7,20 @@
 
 import SwiftUI
 
-// MARK: - View Overlay Gesture Blocker
+// MARK: - View Blocks Gestures
 extension View {
     /// Overlays clear `Rectangle` that blocks gestures if condition is met.
     ///
     ///     var body: some View {
     ///         Button("Lorem Ipsum", action: doSomething)
-    ///             .overlayGestureBlocker(if: !isInteractionEnabled)
+    ///             .blockGestures(!isInteractionEnabled)
     ///     }
     ///
-    public func overlayGestureBlocker(
-        if condition: Bool = true
+    public func blocksGestures(
+        _ flag: Bool = true
     ) -> some View {
         self
-            .if(condition, transform: {
+            .if(flag, transform: {
                 $0
                     .overlay(Color.clear.contentShape(Rectangle()))
             })