+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-### 4.6.0(72)
+### [4.6.0(72)]https://github.com/VakhoKontridze/VCore/releases/tag/4.6.0) — *2023 03 09*
 
 Views